% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvpiler-package.R
\docType{package}
\name{csvpiler-package}
\alias{csvpiler}
\alias{csvpiler-package}
\title{csvpiler: Combine csv files from a folder}
\description{
A lightweight package to combine multiple csv files from a folder into a single csv file.
}
\details{
Functions:
\itemize{
\item\code{pile_csv_min} : Combine csv files into one csv file, no extra column
\item\code{pile_csv()}: Combine csv files into one csv file, and add 2 columns - ID (1st column) and source (last column)
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/FannyOllivier/csvpiler}
  \item Report bugs at \url{https://github.com/FannyOllivier/csvpiler/issues}
}

}
\author{
\strong{Maintainer}: Fanny Ollivier \email{fannyollivier1@gmail.com}

}
