% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pile.R
\name{pile_csv}
\alias{pile_csv}
\title{pile_csv: Pile up csv files with the same columns - adds 2 columns : ID (1st column), source (last column, name of the source file)}
\usage{
pile_csv(dossier, x, sortie = "resultat.csv")
}
\arguments{
\item{dossier}{A folder containing the csv files to pile up, e.g., "C:\\Users\\FichierDonnees".}

\item{x}{A number : number of characters from the csv file names which will be used to form the ID.}

\item{sortie}{A file name for the resulting csv file.}
}
\value{
A csv file with all the rows of the csv files from the folder.
}
\description{
pile_csv: Pile up csv files with the same columns - adds 2 columns : ID (1st column), source (last column, name of the source file)
}
