% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pile.R
\name{pile_csv_min}
\alias{pile_csv_min}
\title{pile_csv_min: Pile up csv files with the same columns - minimal function, no column added}
\usage{
pile_csv_min(dossier, sortie = "resultat.csv")
}
\arguments{
\item{dossier}{A folder containing the csv files to pile up, e.g., "C:\\Users\\FichierDonnees".}

\item{sortie}{A file name for the resulting csv file.}
}
\value{
A csv file with all the rows of the csv files from the folder.
}
\description{
pile_csv_min: Pile up csv files with the same columns - minimal function, no column added
}
